rename .py files
