+++ flowsa/__init__.py
@@ -12,7 +12,7 @@ from flowsa.flowbyfunctions import collapse_fbs_sectors, filter_by_geoscale
 from flowsa.datachecks import check_for_nonetypes_in_sector_col, check_for_negative_flowamounts
 import flowsa.flowbyactivity
 import flowsa.flowbysector
-from flowsa.bibliography_functions import generate_fbs_bibliography
+from flowsa.bibliography import generate_fbs_bibliography
 import logging as log
