+++ flowsa/data_source_scripts/BLS_QCEW.py
@@ -205,8 +205,8 @@ def bls_clean_allocation_fba_w_sec(df_w_sec, **kwargs):
     """
     from flowsa.flowbyfunctions import estimate_suppressed_data, sector_disaggregation, sector_aggregation, \
         flow_by_activity_wsec_mapped_fields, fba_mapped_default_grouping_fields
-    from flowsa.dataclean_functions import add_missing_flow_by_fields
-    from flowsa.dataclean_functions import replace_strings_with_NoneType
+    from flowsa.dataclean import add_missing_flow_by_fields
+    from flowsa.dataclean import replace_strings_with_NoneType
 
     df2 = add_missing_flow_by_fields(df_w_sec, flow_by_activity_wsec_mapped_fields)
     df3 = replace_strings_with_NoneType(df2)
