+++ flowsa/fbs_allocation.py
@@ -9,7 +9,7 @@ from flowsa.datachecks import check_if_losing_sector_data, check_allocation_rati
 from flowsa.flowbyfunctions import fba_activity_fields, fba_mapped_default_grouping_fields, collapse_activity_fields, \
     fbs_activity_fields, sector_aggregation, sector_disaggregation
 from flowsa.mapping import get_fba_allocation_subset
-from flowsa.dataclean_functions import load_map_clean_fba, replace_strings_with_NoneType, \
+from flowsa.dataclean import load_map_clean_fba, replace_strings_with_NoneType, \
     replace_NoneType_with_empty_cells
 
 # import specific functions
